# monitoring: move to OTEL sync gauge