@istio-testing 2024-08-23
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
gencheck_istio | 0bf4a85f3107119182bd2715160375ceef1fd63e | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/52826/gencheck_istio/1826990476117413888) | true | `/test gencheck`
lint_istio | 0bf4a85f3107119182bd2715160375ceef1fd63e | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/52826/lint_istio/1826990484464078848) | true | `/test lint`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes-sigs/prow](https://github.com/kubernetes-sigs/prow/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->