@istio-testing 2024-09-25
In response to a cherrypick label: #52826 failed to apply on top of branch "release-1.23":
```
Applying: add test
Using index info to reconstruct a base tree...
M	pkg/monitoring/monitoring_test.go
Falling back to patching base and 3-way merge...
Auto-merging pkg/monitoring/monitoring_test.go
CONFLICT (content): Merge conflict in pkg/monitoring/monitoring_test.go
error: Failed to merge in the changes.
hint: Use 'git am --show-current-patch=diff' to see the failed patch
Patch failed at 0001 add test
When you have resolved this problem, run "git am --continue".
If you prefer to skip this patch, run "git am --skip" instead.
To restore the original branch and stop patching, run "git am --abort".

```