@istio-testing 2024-09-25
In response to a cherrypick label: new issue created for failed cherrypick: #53292